add: frontend: page de cachorro e gato

+++ backend/src/controllers/clientController.js
@@ -1,10 +1,15 @@
 const Client = require("../services/clientSerevice");
+const { JWT_SECRET } = require("../services/key");
+const jwt = require("jsonwebtoken");
 
 exports.newClient = async (req, res) => {
   try {
+    const token = req.headers.authorization;
+    if (!token) { return res.status(401).json({ message: 'Usuario não autenticado' }); } 
+    const payload = jwt.verify(token, JWT_SECRET);
     const clients = new Client(req.body);
     await clients.register();
-    return res.status(201).json(clients)
+    return res.status(201).json(clients);
   } catch (e) {
     console.log(e);
     return res.status(400).json({ error: e });
@@ -13,6 +18,9 @@ exports.newClient = async (req, res) => {
 
 exports.showClients = async (req, res) => {
   try {
+    const token = req.headers.authorization;
+    if (!token) { return res.status(401).json({ message: 'Usuario não autenticado' }); } 
+    const payload = jwt.verify(token, JWT_SECRET);
     const clients = await Client.findclient();
     return res.status(201).json(clients);
   } catch (e) {
@@ -23,6 +31,9 @@ exports.showClients = async (req, res) => {
 
 exports.showClientById = async (req, res) => {
   try {
+    const token = req.headers.authorization;
+    if (!token) { return res.status(401).json({ message: 'Usuario não autenticado' }); } 
+    const payload = jwt.verify(token, JWT_SECRET);
     const clients = await Client.findClientById(req.params.id);
     return res.status(201).json(clients);
   } catch (e) {
@@ -32,6 +43,9 @@ exports.showClientById = async (req, res) => {
 };
 exports.updateClientById = async (req, res) => {
   try {
+    const token = req.headers.authorization;
+    if (!token) { return res.status(401).json({ message: 'Usuario não autenticado' }); } 
+    const payload = jwt.verify(token, JWT_SECRET);
     const clients = new Client(req.body);
     await clients.edit(req.params.id);
     return res.status(201).json(clients);
@@ -41,14 +55,16 @@ exports.updateClientById = async (req, res) => {
   }
 };
 
-
 exports.deleteClient = async (req, res) => {
-    try{
-        await Client.delete(req.params.id)
-        const clients = await Client.findclient();
-        return res.status(201).json(clients);
-    } catch(e) {
-        console.log(e);
-        return res.status(400).json({ error: e }); 
-    }
-}
+  try {
+    const token = req.headers.authorization;
+    if (!token) { return res.status(401).json({ message: 'Usuario não autenticado' }); } 
+    const payload = jwt.verify(token, JWT_SECRET);
+    await Client.delete(req.params.id);
+    const clients = await Client.findclient();
+    return res.status(201).json(clients);
+  } catch (e) {
+    console.log(e);
+    return res.status(400).json({ error: e });
+  }
+};
